# docs/swagger.yaml
basePath: /api
definitions:
...
  credential.Credential:
    properties:
      description:
        description: Description contains optional user notes about where this credential
          is used.
        example: Email account credentials
        type: string
      id:
        description: ID contains the unique identifier for this credential record.
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      login:
        description: Login contains the username, email, or account identifier (sensitive
          data).
        example: user@example.com
        type: string
      password:
        description: Password contains the plaintext password (highly sensitive, transmitted
          encrypted).
        example: securePassword123
        type: string
      updated_at:
        description: UpdatedAt contains the timestamp when this credential was last
          modified.
        example: "2023-12-01T10:00:00Z"
        type: string
    type: object
  credential.ListResponse:
    properties:
      credentials:
        description: List of credentials
        items:
          $ref: '#/definitions/credential.Credential'
        type: array
    type: object
  credential.PullResponse:
    properties:
      credential:
        allOf:
        - $ref: '#/definitions/credential.Credential'
        description: Credential data
    type: object
  credential.PushRequest:
    properties:
      description:
        description: Optional description
        example: Email account credentials
        type: string
      login:
        description: Login username or email (required)
        example: user@example.com
        type: string
      password:
        description: Password (required)
        example: securePassword123
        type: string
    required:
    - login
    - password
    type: object
  credential.PushResponse:
    properties:
      id:
        description: Created or updated credential ID
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
  response.Error:
    properties:
      messages:
        description: Messages contains one or more error descriptions for the client.
        items:
          type: string
        type: array
    type: object
host: localhost:56789
info:
  contact:
    email: gdyunin@gmail.com
    name: German Dyunin
    url: https://github.com/gdyunin/aegis_vault_keeper
  description: |-
    AegisVaultKeeper is a secure personal data storage service that allows users to store
    and manage sensitive information including credentials, bank cards, notes, and files.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: AegisVaultKeeper API
  version: 0.1.1
paths:
...
  /items/credentials:
    get:
      consumes:
      - application/json
      description: Retrieves all credentials belonging to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Credentials retrieved successfully
          schema:
            $ref: '#/definitions/credential.ListResponse'
        "204":
          description: No credentials found
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: List all credentials
      tags:
      - Credentials
    post:
      consumes:
      - application/json
      description: Creates a new credential or updates an existing one if ID is provided
        in URL path
      parameters:
      - description: Credential data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/credential.PushRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Credential created or updated successfully
          schema:
            $ref: '#/definitions/credential.PushResponse'
        "400":
          description: Bad request - invalid input data
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not found - credential not found for update
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Create or update credential
      tags:
      - Credentials
  /items/credentials/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a specific credential belonging to the authenticated
        user
      parameters:
      - description: Credential ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Credential retrieved successfully
          schema:
            $ref: '#/definitions/credential.PullResponse'
        "400":
          description: Bad request - invalid ID format
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not found - credential not found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get credential by ID
      tags:
      - Credentials
    put:
      consumes:
      - application/json
      description: Creates a new credential or updates an existing one if ID is provided
        in URL path
      parameters:
      - description: Credential ID for update operation
        format: uuid
        in: path
        name: id
        type: string
      - description: Credential data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/credential.PushRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Credential created or updated successfully
          schema:
            $ref: '#/definitions/credential.PushResponse'
        "400":
          description: Bad request - invalid input data
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not found - credential not found for update
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Create or update credential
      tags:
      - Credentials
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication. Use 'Bearer {token}' format.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"